replaced md5 with basename

+++ epigeec/cc/core/genomic_dataset.cc
@@ -42,15 +42,15 @@ std::string GenomicDataset::get_sizes() {
   return sizes;
 }
 
-std::map<std::string, float>  GenomicDataset::Correlate(
+std::map<std::string, double>  GenomicDataset::Correlate(
     GenomicDataset& genomic_dataset,
     std::vector<std::string>& chromosomes) {
-  std::map<std::string, float> results;
+  std::map<std::string, double> results;
   for (const std::string& chr : chromosomes) {
     if (chromosomes_.find(chr) != chromosomes_.end() &&
         genomic_dataset.chromosomes_.find(chr) !=
         genomic_dataset.chromosomes_.end()) {
-      float r;
+      double r;
       try{
         r = chromosomes_.at(chr)->GetPearson(
         *(genomic_dataset.chromosomes().at(chr)));
@@ -59,7 +59,7 @@ std::map<std::string, float>  GenomicDataset::Correlate(
       }
       results.emplace(chr, r);
     } else {
-      results.emplace(chr, std::numeric_limits<float>::quiet_NaN());
+      results.emplace(chr, std::numeric_limits<double>::quiet_NaN());
     }
   }
     return results;
